requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-login==0.6.3
flask-wtf==1.2.2
flask-limiter==3.8.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
apscheduler==3.10.4
telnyx==2.1.3
sendgrid==6.11.0
